requirements.txt
websocket
TA-lib
numpy
binance-clinet